+++ foxes/output/results_writer.py
@@ -1,5 +1,5 @@
 from .output import Output
-
+import foxes.variables as FV
 
 class ResultsWriter(Output):
     """
@@ -22,7 +22,9 @@ class ResultsWriter(Output):
     def __init__(self, farm_results=None, data=None):
 
         if farm_results is not None and data is None:
-            self.data = farm_results.to_dataframe()
+            self.data = farm_results.to_dataframe().reset_index()
+            self.data[FV.TNAME] = farm_results[FV.TNAME].to_numpy()[self.data[FV.TURBINE]]
+            self.data.set_index([FV.STATE, FV.TURBINE], inplace=True)
         elif farm_results is None and data is not None:
             self.data = data
         else:
@@ -34,6 +36,7 @@ class ResultsWriter(Output):
         self,
         file_path,
         variables=None,
+        turbine_names=False,
         verbosity=1,
         **kwargs,
     ):
@@ -49,6 +52,8 @@ class ResultsWriter(Output):
             the keys are the foxes variables and the values
             the column names. If None, then all data will be
             written.
+        turbine_names : bool
+            Use turbine names instead of turbine indices
         verbosity : int
             The verbosity level, 0 = silent
         kwargs : dict, optional
@@ -72,5 +77,8 @@ class ResultsWriter(Output):
                     data = data.rename_axis(index={s: ns})
         else:
             data = self.data[list(variables)]
-
+        
+        if turbine_names:
+            data = data.reset_index().set_index([FV.STATE, FV.TNAME])
+            
         super().write(file_path, data, **kwargs)